add staff logout options

+++ frontend/logout.ts
@@ -3,7 +3,7 @@ import { noop } from "lodash";
 import { API } from "./api";
 import { Session } from "./session";
 
-export const logout = () => {
-  axios.delete(API.current.tokensPath).then(noop, noop);
+export const logout = (keepToken = false) => () => {
+  !keepToken && axios.delete(API.current.tokensPath).then(noop, noop);
   Session.clear();
 };
